% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNonOverlappingTestLabelsOrPoints.R
\name{addLabel}
\alias{addLabel}
\title{Plot a label with optional polygon background}
\usage{
addLabel(x, y, label, cex, col, bg, border, heightPad, widthPad)
}
\arguments{
\item{x}{The X coordinate at which label is to be plotted}

\item{y}{The Y coordinate at which label is to be plotted}

\item{label}{The label to be plotted}

\item{cex}{The number used to scale the size of the label}

\item{col}{The colour of the label to be plotted}

\item{bg}{The colour of the polygon to be plotted. If NULL no polygon plotted}

\item{border}{The colour of the polygon border. If NA, no border plotted}

\item{heightPad}{Multiplyer for label height should added to label to be used to pad height}

\item{widthPad}{Multiplyer for label width should added to label to be used to pad width}
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
