% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{addPlottedLabel}
\alias{addPlottedLabel}
\title{Add the information associated with a text label that has been plotted}
\usage{
addPlottedLabel(x, y, height, width, plottedLabelInfo)
}
\arguments{
\item{x}{X coordinate of point of interest}

\item{y}{Y coodrinate of point of interest}

\item{height}{The height of the label associated with the point of interest}

\item{width}{The width of the label associated with the point of interest}

\item{plottedLabelInfo}{The coordinates and label information about the locations where a label has already plotted}
}
\value{
Returns a list containing information for all the plotted labels, included the one just added
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
