% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarChart.R
\name{calculateAxesCoordinatesAtEachLevel}
\alias{calculateAxesCoordinatesAtEachLevel}
\title{Define the coordinates for the radar level labels}
\usage{
calculateAxesCoordinatesAtEachLevel(axesInfo, nLevels, circles, angle)
}
\arguments{
\item{axesInfo}{A data.frame of X and Y coordinates for end points of each skills axes}

\item{nLevels}{An integer for the number of radar levels}

\item{circles}{A boolean value reporting whether radar chart uses circles or polygons for levels}

\item{angle}{A numeric value representing the angle for the radar levels axes}
}
\value{
Returns a data.frame of the X and Y coordinates of level label coordinates for axes.
}
\description{
Function used by \code{radar()} to calculate locations of radar level labels based on angle
}
\keyword{internal}
