% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{calculateLabelHeightsAndWidths}
\alias{calculateLabelHeightsAndWidths}
\title{Calculate the heights and widths of the labels in the current plotting window}
\usage{
calculateLabelHeightsAndWidths(pointInfo, cex, heightPad, widthPad)
}
\arguments{
\item{pointInfo}{A list storing the coordinates and labels of input points}

\item{cex}{The number used to scale the size of the label and therefore its height and width}

\item{heightPad}{Multiplyer for label height should added to label to be used to pad height}

\item{widthPad}{Multiplyer for label width should added to label to be used to pad width}
}
\value{
Returns a list storing the coordinates, labels, and the heights and widths of the labels, for input points
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
