% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSNPScale.R, R/watermark.R
\name{calculateLocation}
\alias{calculateLocation}
\title{Get the X and Y coordinates of a string coded location on a plot}
\usage{
calculateLocation(location = NULL, axisLimits)

calculateLocation(location = NULL, axisLimits)
}
\arguments{
\item{location}{A character string indicating a location on a plot}

\item{axisLimits}{a numeric vector storing the limits of the axes in the format \code{c(xMin, xMax, yMin, yMax)}.}

\item{location}{a character vector specifying the location for the text to be plotted}
}
\value{
Returns a numeric vector storing the coordinates determined by string location

Returns a numeric vector containing the X and Y coordinates to plot the text at
}
\description{
Function used by \code{addSNPScale()}

Function used by \code{watermark()}
}
\keyword{internal}
