% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{calculatePointShapeSize}
\alias{calculatePointShapeSize}
\title{A function to calculate the size of a point shape in the current plot}
\usage{
calculatePointShapeSize(cex)
}
\arguments{
\item{cex}{A numerical value giving the amount by which plotted symbols WERE magnified relative to their default in the plotted boxplot}
}
\value{
Returns a numerical vector containing the estimated width and height of a point shape
}
\description{
Function used by \code{spreadPoints()}
}
\keyword{internal}
