% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNonOverlappingTestLabelsOrPoints.R
\name{calculatePointSize}
\alias{calculatePointSize}
\title{Calculate the size of a point on the current plot}
\usage{
calculatePointSize(axisLimits, sizeFactor = 1)
}
\arguments{
\item{axisLimits}{The limits of the X and Y axis: (\code{c(xMin, xMax, yMin, yMax)})}

\item{sizeFactor}{A number that increases (values > 1) or decreases (values < 1) the amount of space alloted to each point}
}
\value{
Returns a vector containing the width and height of a point
}
\description{
Function used by \code{addPoints()}
}
\keyword{internal}
