% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{calculateWatermarkCex}
\alias{calculateWatermarkCex}
\title{Calculate the factor to multiple text by so that it fits to plotting window}
\usage{
calculateWatermarkCex(text, axisLimits)
}
\arguments{
\item{text}{a character vector specifying the text to be written}

\item{axisLimits}{a numeric vector storing the limits of the axes in the format \code{c(xMin, xMax, yMin, yMax)}.}
}
\value{
Returns a numeric value to multiple the size of the text to be plotted.
}
\description{
Function used by \code{watermark()}
}
\keyword{internal}
