% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultipleHistograms.R
\name{calculateYAxisLimits}
\alias{calculateYAxisLimits}
\title{Calculate the limits of the Y axis}
\usage{
calculateYAxisLimits(histograms)
}
\arguments{
\item{histograms}{A list of histogram objects}
}
\value{
A vector of length 2 containing 0 and the maximum bin count observed across the histograms
}
\description{
Function used by \code{plotMultipleHistograms()}
}
\keyword{internal}
