% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFasta.R
\name{checkAlignmentClass}
\alias{checkAlignmentClass}
\title{Count the number of different nucleotides at a position in a sequence alignment}
\usage{
checkAlignmentClass(nucleotideAlignment)
}
\arguments{
\item{nucleotideAlignment}{A matrix or a list containing the DNA sequences, or an object of class "alignment", "phyDat", or "DNAbin"}
}
\value{
Returns a nucleotide alignment as matrix of nucleotides in upper case
}
\description{
Function converts the nucleotide alignment to a format that will work with the \code{plotFASTA} function
}
\keyword{internal}
