% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNonOverlappingTestLabelsOrPoints.R
\name{chooseNewLocation}
\alias{chooseNewLocation}
\title{A function to choose (from the alternative locations) a new location for a label to be plotted at}
\usage{
chooseNewLocation(pointInfo, index, alternativeLocations, distances,
  plottedLabelInfo, axisLimits, keepLabelsInside)
}
\arguments{
\item{pointInfo}{A list storing the information for the input points}

\item{index}{The index of the point of interest}

\item{alternativeLocations}{The coordinates of the alternative locations}

\item{distances}{The distances between the alternative locations and the input points}

\item{plottedLabelInfo}{The coordinates and label information about the locations where a label has already plotted}

\item{axisLimits}{The limits of the X and Y axis: (\code{c(xMin, xMax, yMin, yMax)})}

\item{keepLabelsInside}{A logical variable indicating whether the labels shouldn't be plotted outside of plotting region}
}
\value{
Returns the index of the chosen alternative location
}
\description{
Function used by \code{addTextLabels()} and \code{addPoints()}
}
\keyword{internal}
