% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFasta.R
\name{countNucleotidesAtSite}
\alias{countNucleotidesAtSite}
\title{Count the number of different nucleotides at a position in a sequence alignment}
\usage{
countNucleotidesAtSite(position, nucleotideAlignment)
}
\arguments{
\item{position}{A numeric integer indicating the site in the nucleotide alignment to examine}

\item{nucleotideAlignment}{A matrix of DNA sequences as characters}
}
\value{
Returns the number of different nucleotides at each position in the input alignment
}
\description{
Function counts the number of each nucleotide (A, C, G, and T) at position in a nucleotide alignment. Used by \code{plotFASTA} function
}
\keyword{internal}
