% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMultipleHistograms.R
\name{createHistogramObjects}
\alias{createHistogramObjects}
\title{Create, but don't plot, a histogram for each distribution}
\usage{
createHistogramObjects(distributions, breaks)
}
\arguments{
\item{distributions}{A list containing multiple distributions - each represented as a numeric vector}

\item{breaks}{A vector defining the start and end of bins}
}
\value{
Returns a list of histogram objects
}
\description{
Function used by \code{plotMultipleHistograms()}
}
\keyword{internal}
