% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMultipleHistograms.R
\name{defineBreaks}
\alias{defineBreaks}
\title{Define the breaks to be used in histograms}
\usage{
defineBreaks(xLimits, nBins)
}
\arguments{
\item{xLimits}{A vector of length 2 representing the range of the distributions}

\item{nBins}{A numeric value indicating the number of bins to define}
}
\value{
Returns a vector defining the start and end of bins
}
\description{
Function used by \code{plotMultipleHistograms()}
}
\keyword{internal}
