% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Calculate the euclidean distance between two points}
\usage{
euclideanDistance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{The X coordinate of the first point}

\item{y1}{The Y coordinate of the first point}

\item{x2}{The X coordinate of the second point}

\item{y2}{The Y coordinate of the second point}
}
\value{
Returns the distance between the points provided
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
