% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarChart.R
\name{generateEquiDistantPointsOnCircle}
\alias{generateEquiDistantPointsOnCircle}
\title{Generate equidistant points on a circle}
\usage{
generateEquiDistantPointsOnCircle(nPoints, radius = 1, origin = c(0, 0))
}
\arguments{
\item{nPoints}{Integer for the number of points to position on circle}

\item{radius}{Numeric value for the radius of the circle. Default value is 1.}

\item{origin}{Numeric vector of length 2 that provides the origin coordinates of the circle. Defaults to \code{c(0,0)}}
}
\value{
Returns a data.frame of the X and Y coordinates of points around the circle.
}
\description{
Function used by \code{radarChart()} to create the locations for the skills axes
}
\keyword{internal}
