% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarChart.R
\name{generatePointsAlongPolygon}
\alias{generatePointsAlongPolygon}
\title{Places set number of points along the edges of a polygon.}
\usage{
generatePointsAlongPolygon(coords, nPoints)
}
\arguments{
\item{coords}{A data.frame of X and Y coordinates for end points of each skills axes}

\item{nPoints}{The number of points to position on polygon, must be at least twice the number of skills axes}
}
\value{
Returns a data.frame of the X and Y coordinates of points along polygon edges. Note includes polygon coordinates.
}
\description{
Function used by \code{calculateAxesCoordinatesAtEachLevel()} to create possible locations for the end of the levels axes
}
\keyword{internal}
