% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{identifyValuesInBins}
\alias{identifyValuesInBins}
\title{A function used to identify the indices of the values falling into each bin}
\usage{
identifyValuesInBins(bins)
}
\arguments{
\item{bins}{A categorical vector detailing the start and end of each bin each value fell in. Defined by \code{cut()} function in \code{spreadPoints()}}
}
\value{
Returns a list detailing the indices of the values that fell in each bin. Keys for list are the bins.
}
\description{
Function used by \code{spreadPoints()}
}
\keyword{internal}
