% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{identifyWhiskersAndOutliers}
\alias{identifyWhiskersAndOutliers}
\title{A function to identify the locations of the whiskers in a boxplot and identify outlier values outside the whiskers}
\usage{
identifyWhiskersAndOutliers(values, range)
}
\arguments{
\item{values}{A vector of numerical values}

\item{range}{Numerical value used to determine outliers. Outliers are defined as those outside \code{range} * InterQuartileRange}
}
\value{
Returns a list containing two vectors: "Upper" and "Lower". THe first value in these vectors is the whisker location. The remaining values are outliers.
}
\description{
Function used by \code{spreadPoints()}
}
\keyword{internal}
