% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{outsidePlot}
\alias{outsidePlot}
\title{Checks whether adding a label at the current point will ending up being outside of the plotting window}
\usage{
outsidePlot(x, y, height, width, axisLimits)
}
\arguments{
\item{x}{X coordinate of point of interest}

\item{y}{Y coodrinate of point of interest}

\item{height}{The height of the label associated with the point of interest}

\item{width}{The width of the label associated with the point of interest}

\item{axisLimits}{The limits of the X and Y axis: (\code{c(xMin, xMax, yMin, yMax)})}
}
\value{
Returns a logical variable to indicate whether the point of interest was too close to any plotted labels
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
