% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNonOverlappingTestLabelsOrPoints.R
\name{overlapsWithPlottedPoints}
\alias{overlapsWithPlottedPoints}
\title{Checks whether a point is too close to any of the plotted points}
\usage{
overlapsWithPlottedPoints(x, y, height, width, pointInfo)
}
\arguments{
\item{x}{X coordinate of point of interest}

\item{y}{Y coodrinate of point of interest}

\item{height}{The height of the label associated with the point of interest}

\item{width}{The width of the label associated with the point of interest}

\item{pointInfo}{A list storing the information for the input points - that have been plotted}
}
\value{
Returns a logical variable to indicate whether the point of interest was too close to any plotted points
}
\description{
Function used by \code{addTextLabels()} and \code{addPoints()}
}
\keyword{internal}
