% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{plotBins}
\alias{plotBins}
\title{A function used in development to see the size of the bins used to spread points that fall within them}
\usage{
plotBins(bins, col)
}
\arguments{
\item{bins}{A categorical vector detailing the start and end of each bin each value fell in Defined by \code{cut()} function in \code{spreadPoints()}}

\item{col}{The colour of the horizontal lines, which define the bin start and ends, to be plotted}
}
\description{
Function used in development by \code{spreadPoints()}
}
\keyword{internal}
