% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFasta.R
\name{plotFASTA}
\alias{plotFASTA}
\title{Plot the different nucleotides present in an alignment of nucleotide sequences}
\usage{
plotFASTA(nucleotideAlignment, pdfFileName = NULL, pdfWidth = 14,
  pdfHeight = 7, labelSpace = 1, lineForSequenceNames = 0,
  sequenceLabelCex = 0.5, xLabCex = 1, xTicksCex = 1,
  legendCex = 1, showNucleotide = FALSE)
}
\arguments{
\item{nucleotideAlignment}{A matrix or a list containing the DNA sequences, or an object of class "alignment", "phyDat", or "DNAbin"}

\item{pdfFileName}{A character string providing the name of a PDF to send plot into. Default value is NULL (plot not sent to file)}

\item{pdfWidth}{The width of the graphics region in inches. Default value is 14.}

\item{pdfHeight}{The height of the graphics region in inches. Default value is 7.}

\item{labelSpace}{Numeric multiplier to change the size of the space alotted to the sequence labels}

\item{lineForSequenceNames}{The number of lines into the margin at which the sequence names will be plotted. Default value is 0.}

\item{sequenceLabelCex}{Numeric multiplier to change the size of the sequence labels. Default value is 0.5.}

\item{xLabCex}{Numeric multiplier to change size of X axis label. Default value is 1.}

\item{xTicksCex}{Numeric multipler to change size of X axis tick labels. Default value is 1.}

\item{legendCex}{Numeric multiplier to change size of legend (points and text). Default value is 1.}

\item{showNucleotide}{Boolean flag to decide whether or not to plot nucleotide characters onto figure. Defaults to off.}
}
\description{
Function uses coloured polygons to illustrate which nucleotides are present in each sequence at each position
}
\examples{
# Create an example dataset
nSequences <- 50
nSites <- 100
alignment <- matrix(sample(c("A", "C", "G", "T"), nSequences*nSites, replace=TRUE),
                    nrow=nSequences, ncol=nSites)
rownames(alignment) <- paste("Seq-", 1:nSequences)

# Plot the FASTA
plotFASTA(alignment, xTicksCex = 1.5, xLabCex=1.5, sequenceLabelCex=0.4)
}
\keyword{DNAbin}
\keyword{alignment}
\keyword{nucleotide}
\keyword{plot}
