% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{A progress bar function to run within a \code{for} loop}
\usage{
progress(i, n)
}
\arguments{
\item{i}{A numeric value indicating the current iteration of the containing for loop}

\item{n}{A numeric value indicating the total number of iterations of the containing for loop}
}
\description{
This function should be ran within a \code{for} loop, it produces a progress bar to indicate how many iterations have passed
}
\examples{
# Set the number of iterations of the for loop
n <- 1000

for(i in 1:n){
 
  # Sleep for a tenth of a second
  Sys.sleep(0.01)
  
  # Update the progress bar
  progress(i, n)
}
}
\keyword{bar}
\keyword{for}
\keyword{progress}
