% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultipleHistograms.R
\name{removeNAValues}
\alias{removeNAValues}
\title{Remove \code{NA} values from each distribution}
\usage{
removeNAValues(distributions)
}
\arguments{
\item{distributions}{A list containing multiple distributions - each represented as a numeric vector}
}
\value{
Returns a list containing multiple distributions without any NA values
}
\description{
Function used by \code{plotMultipleHistograms()}
}
\keyword{internal}
