% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAlpha.R, R/spreadPoints.R
\name{setAlpha}
\alias{setAlpha}
\title{A progress bar function to run within a \code{for} loop}
\usage{
setAlpha(colour, alpha)

setAlpha(colour, alpha)
}
\arguments{
\item{colour}{A colour}

\item{alpha}{A numeric value (between 0 and 1), indicating the level of transparency to used}

\item{colours}{A vector of colours defined by strings of characters}

\item{alpha}{The transparency (0=transparent, 1=opaque) to be assigned to the input colour}
}
\value{
Returns a colour code representing the input colour with the transparency provided
}
\description{
A function to change the alpha value (transparency) of colours that are defined as strings.

Function used by \code{spreadPoints()}
}
\examples{
# Define the coordinates of three points
x <- c(0.2, 0.5, 0.8)
y <- c(0.5, 0.5, 0.5)

# Define the colours for the points
colours <- c("green", "blue", "red")

# Plot the points - set the transparency
plot(x, y, bty="n", xlab="X", ylab="Y", las=1, pch=19, cex=30, xlim=c(0,1), ylim=c(0,1),
     col=setAlpha(colours, alpha=0.5))
}
\keyword{alpha}
\keyword{colour}
\keyword{internal}
\keyword{transparency}
