% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNonOverlappingTestLabelsOrPoints.R,
%   R/spreadPoints.R
\name{setOption}
\alias{setOption}
\title{A function to assign value if multiple options are available, can recycle if index is > number of options available}
\usage{
setOption(options, index)

setOption(options, index)
}
\arguments{
\item{index}{The current index of a label}

\item{colours}{A single option of vector of options}

\item{colours}{A single option of vector of options}

\item{index}{The current index of a label}
}
\value{
Returns the selected option based upon the index provided

Returns the selected option based upon the index provided
}
\description{
Function used by \code{addTextLabels()} and \code{addPoints()}

Function used by \code{addTextLabels()}
}
\keyword{internal}
