% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{spreadPoints}
\alias{spreadPoints}
\title{Add spreaded, to avoid overlap, points to a boxplot}
\usage{
spreadPoints(
  values,
  position,
  pointCex = 1,
  col = "black",
  pch = 19,
  alpha = 0.5,
  plotBins = FALSE,
  plotOutliers = FALSE,
  range = 1.5,
  horiz = FALSE,
  fitToBoxWidth = TRUE,
  xpd = FALSE,
  widthCex = 1
)
}
\arguments{
\item{values}{A vector of numerical values}

\item{position}{A numerical value indicating the position on the x-axis where values are to be plotted. Boxplots use integers: 1,2,3,...}

\item{pointCex}{A numerical value giving the amount by which plotted symbols should be magnified relative to their default. Default value is 1}

\item{col}{The colour of the points to be plotted. Defaults to black}

\item{pch}{The shape of the points to be plotted}

\item{alpha}{The transparency (0=transparent, 1=opaque). Default value is 0.5}

\item{plotBins}{Boolean parameter indicating whether the bins used to spread to points should be plotted as horizontal lines}

\item{plotOutliers}{Boolean parameter indicating whether to plot outliers. Outliers are defined as those outside \code{range} * InterQuartileRange. Defaults to FALSE}

\item{range}{Numerical value used to determine outliers. Default value is 1.5 - same as used by boxplot function}

\item{horiz}{Boolean parameter indicating whether boxplot was plotted horizontally. Default value is FALSE}

\item{fitToBoxWidth}{Boolean parameter indicating whether the points are to spread only within the width of the box. Default value is TRUE}

\item{xpd}{A Boolean value or NA. If FALSE, all plotting is clipped to the plot region, if TRUE, all plotting is clipped to the figure region, and if NA, all plotting is clipped to the device region}

\item{widthCex}{A numerical value giving the amount by which amount the points are spread out should be magnified relative to their default. Default value is 1}
}
\description{
This function is similar to \code{stripchart()} function except it spreads points along an axis in a deterministic rather than random manner
}
\examples{
# Generate some example points - drawn from exponential distribution
values <- rexp(n=50, rate=2)

# Plot a boxplot
boxplot(values, xlab="",  ylab="", frame=FALSE, las=1, pch=19, outcol=rgb(1,0,0, 0.5),
        horizontal=FALSE)

# Plot the points spread along the X axis
spreadPoints(values, position=1)
}
\keyword{boxplot}
\keyword{points}
\keyword{spread}
