% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{spreadPointsMultiple}
\alias{spreadPointsMultiple}
\title{Add spreaded, to avoid overlap, points to a boxplot containing multiple boxes drawn using data from dataframe}
\usage{
spreadPointsMultiple(data, responseColumn, categoriesColumn,
  pointCex = 1, col = "black", pch = 19, alpha = 0.5,
  plotBins = FALSE, plotOutliers = FALSE, range = 1.5,
  horiz = FALSE, fitToBoxWidth = TRUE, xpd = FALSE, widthCex = 1)
}
\arguments{
\item{data}{A dataframe object containing the data to be plotted}

\item{responseColumn}{The name or index of the response data (Y)}

\item{categoriesColumn}{The name or index of the column that categorises the response data}

\item{pointCex}{A numerical value giving the amount by which plotted symbols should be magnified relative to their default. Default value is 1}

\item{col}{The colour of the points to be plotted. Defaults to black}

\item{pch}{The shape of the points to be plotted}

\item{alpha}{The transparency (0=transparent, 1=opaque). Default value is 0.5}

\item{plotBins}{Boolean parameter indicating whether the bins used to spread to points should be plotted as horizontal lines}

\item{plotOutliers}{Boolean parameter indicating whether to plot outliers. Outliers are defined as those outside \code{range} * InterQuartileRange. Defaults to FALSE}

\item{range}{Numerical value used to determine outliers. Default value is 1.5 - same as used by boxplot function}

\item{horiz}{Boolean parameter indicating whether boxplot was plotted horizontally. Default value is FALSE}

\item{fitToBoxWidth}{Boolean parameter indicating whether the points are to spread only within the width of the box. Default value is TRUE}

\item{xpd}{A Boolean value or NA. If FALSE, all plotting is clipped to the plot region, if TRUE, all plotting is clipped to the figure region, and if NA, all plotting is clipped to the device region}

\item{widthCex}{A numerical value giving the amount by which amount the points are spread out should be magnified relative to their default. Default value is 1}
}
\description{
This function is similar to \code{stripchart()} function except it spreads points along an axis in a deterministic rather than random manner
}
\examples{
# Generate some example points - drawn from normal distribution and randomly assign them to categories
randomSamples <- data.frame(Values = rnorm(500), Category = sample(c('A', 'B', 'C', 'D', 'E'), size=500, replace=TRUE))

# Plot a boxplot of the samples from the normal distribution versus there categories - multiple boxplots
boxplot(Values ~ Category, data = randomSamples, lwd = 2)

# Plot the points for each category spread along the X axis
spreadPointsMultiple(data=randomSamples, responseColumn="Values", categoriesColumn="Category")
}
\keyword{boxplot}
\keyword{points}
\keyword{spread}
