% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{testPchSizeCalculator}
\alias{testPchSizeCalculator}
\title{A function used in development to see how well the size of a point shape was calculated.}
\usage{
testPchSizeCalculator(ptSize, coords)
}
\arguments{
\item{ptSize}{A numerical vector containing the estimated width and height of a point shape}

\item{coords}{A numerical vector containing the X and Y coordinates to plot a point}
}
\description{
Function used in development by \code{spreadPoints()}
}
\keyword{internal}
