% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNonOverlappingTestLabelsOrPoints.R
\name{tooClose}
\alias{tooClose}
\title{Checks whether a point is too close to any of the plotted labels}
\usage{
tooClose(x, y, height, width, plottedLabelInfo)
}
\arguments{
\item{x}{X coordinate of point of interest}

\item{y}{Y coodrinate of point of interest}

\item{height}{The height of the label associated with the point of interest}

\item{width}{The width of the label associated with the point of interest}

\item{plottedLabelInfo}{The coordinates and label information about the locations where a label has already plotted}
}
\value{
Returns a logical variable to indicate whether the point of interest was too close to any plotted labels
}
\description{
Function used by \code{addTextLabels()} and \code{addPoints()}
}
\keyword{internal}
