% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Add a watermark to an existing plot}
\usage{
watermark(text, location = NULL, x = NULL, y = NULL, col = "black",
  alpha = 0.5, srt = 45, cex = 1, ...)
}
\arguments{
\item{text}{a character vector specifying the text to be written}

\item{location}{an optional character vector specifying the location for the text to be plotted}

\item{x}{optional x co-ordinate to be used to position the text}

\item{y}{optional y co-ordinate to be used to position the text}

\item{col}{the colour of the text to be plotted. Defaults to black.}

\item{alpha}{the transparency (0=translucent and 1=opaque) of the text to be plotted. Defaults to 0.5.}

\item{srt}{the number of degrees from horizontal to set the angle of the text}

\item{cex}{numeric character expansion factor. By default, text will be stretched to fit plotting window.}
}
\description{
This function should be run following a plot. It will add a transparent text label to act as a watermark.
}
\examples{
# Generate some example data
x <- runif(10000)
y <- runif(10000)

# Plot the points
plot(x, y, las=1, bty="n", pch=19, col=rgb(0,0,0, 0.1))

# Add a watermark
watermark("CONFIDENTIAL", col="red")
}
\keyword{plot}
\keyword{text}
\keyword{watermark}
