% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSNPScale.R
\name{addSNPScale}
\alias{addSNPScale}
\title{Add a SNP scale to phylogeny}
\usage{
addSNPScale(x = NULL, y = NULL, position = NULL, size = 1,
  lineWidth = 1)
}
\arguments{
\item{x}{An optional numeric X coordinate for the scale}

\item{y}{An optional numeric Y coordinate for the scale}

\item{postion}{An optional character vector detailing the location of the plot (top, middle, left, right, etc.)}
}
\description{
Function that adds a SNP scale onto a plotted phylogeny
}
\keyword{nucleotide}
\keyword{phylogeny}
\keyword{scale}
\keyword{snp}
