% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSNPScale.R
\name{addSNPScale}
\alias{addSNPScale}
\title{Add a SNP scale to phylogeny}
\usage{
addSNPScale(x = NULL, y = NULL, position = NULL, size = 1,
  lineWidth = 1, cex = 1, offset = 0.5)
}
\arguments{
\item{x}{An optional numeric X coordinate for the scale}

\item{y}{An optional numeric Y coordinate for the scale}

\item{lineWidth}{A numeric scaling factor to change the width of the scale line. Defaults to 1. Equivalent to \code{lwd}}

\item{cex}{A numeric scaling factor to chancge the size of the scale label. Defaults to 1.}

\item{offset}{This value controls the distance (‘offset’) of the text label from the scale in fractions of a character width. Defaults to 0.5.}

\item{postion}{An optional character vector detailing the location of the plot (top, middle, left, right, etc.)}
}
\description{
Function that adds a SNP scale onto a plotted phylogeny
}
\keyword{nucleotide}
\keyword{phylogeny}
\keyword{scale}
\keyword{snp}
