% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{calculateLocation}
\alias{calculateLocation}
\title{Get/calculate the location of the text to be plotted. Either as centre of plotting window or specified text location.}
\usage{
calculateLocation(location = NULL, axisLimits)
}
\arguments{
\item{location}{a character vector specifying the location for the text to be plotted}

\item{axisLimits}{a numeric vector storing the limits of the axes in the format \code{c(xMin, xMax, yMin, yMax)}.}
}
\value{
Returns a numeric vector containing the X and Y coordinates to plot the text at
}
\description{
Function used by \code{watermark()}
}
\keyword{internal}
