% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{generateAlternativeLocations}
\alias{generateAlternativeLocations}
\title{Generate a set of alternative locations where labels can be plotted if they overlap with another label}
\usage{
generateAlternativeLocations(axisLimits)
}
\arguments{
\item{axisLimits}{The limits of the X and Y axis: (\code{c(xMin, xMax, yMin, yMax)})}
}
\value{
Returns a list containing the coordinates of the alternative locations
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
