% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadPoints.R
\name{setAlpha}
\alias{setAlpha}
\title{A function used to set the transparency of a colour}
\usage{
setAlpha(colour, alpha)
}
\arguments{
\item{colour}{A colour}

\item{alpha}{The transparency (0=transparent, 1=opaque) to be assigned to the input colour}
}
\value{
Returns a colour code representing the input colour with the transparency provided
}
\description{
Function used by \code{spreadPoints()}
}
\keyword{internal}
